modularized navbar
